refactor(logWidget): log widget uses a template
